TrainingServer/requirements.txt
ultralytics
grpcio
grpcio-tools